+++ src/tests/index.test.ts
@@ -5,6 +5,7 @@ import { before, after, describe, it } from 'node:test';
 import stringify from 'qs/lib/stringify';
 
 import { readChunks, startServer } from './server';
+import pkg from '../../package.json';
 import { merge } from '../index';
 import { XiorTimeoutError } from '../utils';
 import { xior } from '../xior';
@@ -23,6 +24,10 @@ after(async () => {
 describe('xior tests', () => {
   describe('should work with methods GET/POST/PUT/PATCH/DELETE/HEAD/OPTIONS', () => {
     const xiorInstance = xior.create({ baseURL });
+
+    it('xior.VERSION should be same with package.json', async () => {
+      assert.strictEqual(xior.VERSION, pkg.version);
+    });
 
     it('GET should work', async () => {
       const { data } = await xiorInstance.get<{
@@ -55,10 +60,12 @@ describe('xior tests', () => {
         method: string;
         body: Record<string, any>;
         query: Record<string, any>;
-      }>('/delete', { params: { a: 1, b: 2 } });
+      }>('/delete', { params: { a: 1, b: 2 }, data: { c: 3 } });
       assert.strictEqual(data.method, 'delete');
       assert.strictEqual(data.query.a, '1');
       assert.strictEqual(data.query.b, '2');
+      assert.strictEqual(data.query.c, '3');
+      assert.strictEqual(Object.keys(data.body).length, 0);
     });
 
     it('PUT should work', async () => {