+++ src/utils.ts
@@ -52,6 +52,18 @@ export function isAbsoluteURL(url: string) {
   return /^([a-z][a-z\d+\-.]*:)?\/\//i.test(url);
 }
 
+/**
+ * joinPath('/', '/') -> '/'
+ * joinPath('/a/', '/b') -> '/a/b'
+ * joinPath('/a', '/b') -> '/a/b'
+ */
+export function joinPath(path1: string, path2: string) {
+  if (!path2) return path1;
+  return (
+    (path1.endsWith('/') ? path1 : path1 + '/') + (path2.startsWith('/') ? path2.slice(1) : path2)
+  );
+}
+
 export class XiorError extends Error {
   request?: XiorRequestConfig;
   config?: XiorRequestConfig;