+++ src/xior.ts
@@ -1,4 +1,4 @@
-import { defaultRequestInterceptor } from './interceptors';
+import defaultRequestInterceptor from './interceptors';
 import type {
   XiorInterceptorRequestConfig,
   XiorPlugin,
@@ -12,6 +12,7 @@ import {
   isAbsoluteURL,
   XiorError,
   merge,
+  joinPath,
 } from './utils';
 
 const supportAbortController = typeof AbortController !== 'undefined';
@@ -20,6 +21,7 @@ export class xior {
   static create(options?: XiorRequestConfig) {
     return new xior(options);
   }
+  static VERSION = '0.2.1';
 
   config?: XiorRequestConfig;
   defaults: XiorInterceptorRequestConfig;
@@ -152,7 +154,16 @@ export class xior {
   }
 
   private async handlerFetch<T>(requestConfig: XiorRequestConfig): Promise<XiorResponse<T>> {
-    const { url, method, headers, timeout, signal: reqSignal, data, ...rest } = requestConfig;
+    const {
+      url,
+      method,
+      headers,
+      timeout,
+      signal: reqSignal,
+      data,
+      _data,
+      ...rest
+    } = requestConfig;
 
     /** timeout */
     let signal: AbortSignal;
@@ -177,14 +188,11 @@ export class xior {
 
     let finalURL = requestConfig._url || url || '';
     if (requestConfig.baseURL && !isAbsoluteURL(finalURL)) {
-      const baseURL = requestConfig.baseURL.endsWith('/')
-        ? requestConfig.baseURL
-        : requestConfig.baseURL + '/';
-      finalURL = baseURL + (finalURL.startsWith('/') ? finalURL.slice(1) : finalURL);
+      finalURL = joinPath(requestConfig.baseURL, finalURL);
     }
 
     const response = await fetch(finalURL, {
-      body: ['HEAD', 'GET'].includes(requestConfig.method || 'GET') ? undefined : data,
+      body: ['HEAD', 'GET', 'DELETE'].includes(requestConfig.method || 'GET') ? undefined : _data,
       ...rest,
       signal,
       method,
@@ -194,6 +202,13 @@ export class xior {
     if (timer) clearTimeout(timer);
     (signal as ClearableSignal)?.clear?.();
 
+    const commonRes = {
+      response,
+      config: requestConfig as XiorInterceptorRequestConfig,
+      status: response.status,
+      statusText: response.statusText,
+      headers: response.headers,
+    };
     if (!response.ok) {
       let data: any = undefined;
       try {
@@ -207,17 +222,14 @@ export class xior {
         !response.status ? `Network error` : `Request failed with status code ${response.status}`,
         requestConfig,
         {
-          response,
           data,
-          config: requestConfig as XiorInterceptorRequestConfig,
-          status: response.status,
-          statusText: response.statusText,
-          headers: response.headers,
+          ...commonRes,
         }
       );
       for (const item of this.responseInterceptors) {
         if (item.onRejected) {
-          await item.onRejected(error);
+          const res = await item.onRejected(error);
+          if (res?.response?.ok) return res;
         }
       }
       throw error;
@@ -227,42 +239,24 @@ export class xior {
       return {
         data: undefined as T,
         request: requestConfig,
-        config: requestConfig as XiorInterceptorRequestConfig,
-        response,
-        headers: response.headers,
-        status: response.status,
-        statusText: response.statusText,
+        ...commonRes,
       };
     }
 
-    if (
-      !requestConfig.responseType ||
-      requestConfig.responseType === 'json' ||
-      requestConfig.responseType === 'text'
-    ) {
-      let data: {
-        data: T;
-      };
+    const { responseType } = requestConfig;
+    if (!responseType || responseType === 'json' || responseType === 'text') {
+      let data: any;
       try {
-        data = {
-          data: (await response.text()) as T,
-        };
-        try {
-          if (data.data && requestConfig.responseType !== 'text') {
-            data.data = JSON.parse(data.data as string);
-          }
-          // eslint-disable-next-line @typescript-eslint/no-unused-vars
-        } catch (e) {
-          //
+        data = (await response.text()) as T;
+        if (data && responseType !== 'text') {
+          data = JSON.parse(data as string);
         }
         // eslint-disable-next-line @typescript-eslint/no-unused-vars
       } catch (e) {
-        data = {
-          data: undefined as T,
-        };
+        //
       }
       let responseObj = {
-        data: data.data,
+        data: data as T,
         request: requestConfig as XiorInterceptorRequestConfig,
         response,
       };
@@ -273,22 +267,14 @@ export class xior {
       return {
         data: responseObj.data,
         request: requestConfig,
-        config: requestConfig as XiorInterceptorRequestConfig,
-        response,
-        headers: response.headers,
-        status: response.status,
-        statusText: response.statusText,
+        ...commonRes,
       };
     }
 
     return {
       data: undefined as T,
       request: requestConfig,
-      config: requestConfig as XiorInterceptorRequestConfig,
-      response,
-      headers: response.headers,
-      status: response.status,
-      statusText: response.statusText,
+      ...commonRes,
     };
   }